+++ src/main/java/cz/aegis/corebanking/repository/TransactionRepository.java
@@ -1,0 +1,8 @@
+package cz.aegis.corebanking.repository;
+
+import cz.aegis.corebanking.entity.Transaction;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TransactionRepository
+        extends JpaRepository<Transaction, Long> {
+}