+++ docs/CHANGELOG.md
@@ -43,4 +43,5 @@
 - Vytvoření Transfer entity a relací sourceAccount, targetAccount -> aktualizace dokumentace (PROJECT_CONTEXT.md)
 - Vytvoření TransferRepository a první integrační test pro entitu Transfer -> aktualizace dokumentace (PROJECT_CONTEXT.md)
 - Vytvoření Transaction entity -> aktualizace dokumentace (PROJECT_CONTEXT.md)
+- Oprava Transaction entity mapování v databázovém schámatu, přidání chybějícího pole, oprava odkazujících atributů (tx_id, tx_type, tx_date mapování)
 -